requirements.txt
requests==2.31.0
tweepy==4.14.0
python-telegram-bot==20.3
asyncio~=3.4.3
logging~=0.4.9.6
python-dotenv==1.0.0
